Rename citeproc.string to prevent shadowing stdlib
